+++ rust/src/router/consumer.rs
@@ -408,7 +408,7 @@ pub struct ConsumerStat {
     pub fir_count: u64,
     pub packet_count: u64,
     pub byte_count: u64,
-    pub bitrate: u32,
+    pub bitrate: u64,
     pub round_trip_time: Option<f32>,
     pub rtx_packets_discarded: Option<u64>,
     pub score: u8,